CorrecciondeControladores
Se agregaron recursos para cada modelo y en cada return de los metodos de los controladores se agrego el uso del recurso para formatear la salida del modelo. Se arreglo el problema del storage y el determinado de la ruta de acceso. Se corrigio el muestreo de la paginacion en .json para Collections. Se agrego enum de HTTP_INTERNAL_SERVER_ERROR.

+++ database/seeders/DatabaseSeeder.php
@@ -15,8 +15,11 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         // Elimina y crea el directorio de portadas
-        Storage::deleteDirectory('public/portadas');
-        Storage::makeDirectory('public/portadas');
+        if (Storage::exists('public/portadas')){
+            Storage::deleteDirectory('public/portadas');
+        } else {
+            Storage::makeDirectory('public/portadas');
+        }
 
         // Crea un usuario de prueba
         User::factory()->create([
